requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Assets==0.9
Flask-Babel==0.9
Flask-Login==0.2.10
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.18
SQLAlchemy==0.9.3
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.3
blinker==1.3
coverage==3.7.1
cssmin==0.2.0
docopt==0.6.1
factory-boy==2.3.1
gunicorn==18.0
itsdangerous==0.23
mock==1.0.1
nose==1.3.0
pep8==1.4.6
psycopg2==2.5.2
pytz==2013.9
speaklater==1.3
webassets==0.9
wsgiref==0.1.2
